Refactor UI for polymorphism when using Cli and Gui
